fix(lint): split ui/routines/filter_tests.rs under the 600-line gate
filter_tests.rs (650 lines) exceeded the lowered 600-line pre-push gate.
Moves the RoutineStatusFacet/AgentFacet/RepositoryFacet/TagFacet/
RoutineMachineFacet codec roundtrip tests into a new sibling
filter_facet_codec_tests.rs, mirroring the existing filter_health_tests.rs
mod-declaration style in filter.rs. No behavior or test-body changes.
